+++ src/event/window.rs
@@ -1,7 +1,7 @@
 use crate::Event;
 
 pub enum WindowEvent {
-    Resize { width: u32, height: u32 },
+    Resize { width: f64, height: f64 },
 }
 
 impl Event for WindowEvent {}
